TST: move binom_conf_interval tests to the test_utils file. Also remove binom_conf_interval and incomplete hypergeom_conf_interval from core.py.
