+++ permute/core.py
@@ -46,117 +46,6 @@ def corr(x, y, reps=10**4, seed=None):
     return tst, left_pv, right_pv, two_sided_pv, sims
 
 
-def binom_conf_interval(n, x, cl=0.975, alternative="two-sided", p=None,
-                        **kwargs):
-    """
-    Compute a confidence interval for a binomial p, the probability of success in each trial.
-
-    Parameters
-    ----------
-    n : int
-        The number of Bernoulli trials.
-    x : int
-        The number of successes.
-    cl : float in (0, 1)
-        The desired confidence level.
-    alternative : {"two-sided", "lower", "upper"}
-        Indicates the alternative hypothesis.
-    p : float in (0, 1)
-        Starting point in search for confidence bounds for probability of success in each trial.
-    kwargs : dict
-        Key word arguments
-
-    Returns
-    -------
-    tuple
-        lower and upper confidence level with coverage (approximately)
-        1-alpha.
-
-    Notes
-    -----
-    xtol : float
-        Tolerance
-    rtol : float
-        Tolerance
-    maxiter : int
-        Maximum number of iterations.
-    """
-    assert alternative in ("two-sided", "lower", "upper")
-
-    if p is None:
-        p = x / n
-    ci_low = 0.0
-    ci_upp = 1.0
-
-    if alternative == 'two-sided':
-        cl = 1 - (1-cl)/2
-
-    if alternative != "upper" and x > 0:
-        f = lambda q: cl - binom.cdf(x - 1, n, q)
-        ci_low = brentq(f, 0.0, p, *kwargs)
-    if alternative != "lower" and x < n:
-        f = lambda q: binom.cdf(x, n, q) - (1 - cl)
-        ci_upp = brentq(f, 1.0, p, *kwargs)
-
-    return ci_low, ci_upp
-
-def hypergeom_conf_interval(n, x, N, cl=0.975, alternative="two-sided", G=None,
-                        **kwargs):
-    """
-    Compute a confidence interval for a hypergeometric G, the number of good objects in the population.
-
-    Parameters
-    ----------
-    n : int
-        The number of Bernoulli trials.
-    x : int
-        The number of "good" objects in the sample.
-    N : int
-        The number of objects in the population
-    cl : float in (0, 1)
-        The desired confidence level.
-    alternative : {"two-sided", "lower", "upper"}
-        Indicates the alternative hypothesis.
-    G : int in [0, N]
-        Starting point in search for confidence bounds for hypergeometric G.
-    kwargs : dict
-        Key word arguments
-
-    Returns
-    -------
-    tuple
-        lower and upper confidence level with coverage (at least)
-        1-alpha.
-
-    Notes
-    -----
-    xtol : float
-        Tolerance
-    rtol : float
-        Tolerance
-    maxiter : int
-        Maximum number of iterations.
-    """
-    assert alternative in ("two-sided", "lower", "upper")
-
-    if G is None:
-        G = (x / n)*N
-    ci_low = 0.0
-    ci_upp = N
-
-    if alternative == 'two-sided':
-        cl = 1 - (1-cl)/2
-
-    if alternative != "upper" and x > 0:
-        f = lambda q: cl - binom.cdf(x - 1, n, q)
-        ci_low = brentq(f, 0.0, p, *kwargs)
-    if alternative != "lower" and x < n:
-        f = lambda q: binom.cdf(x, n, q) - (1 - cl)
-        ci_upp = brentq(f, 1.0, p, *kwargs)
-
-    return ci_low, ci_upp
-
-
 def two_sample(x, y, reps=10**5, stat='mean', alternative="greater",
                keep_dist=False, interval=False, level=0.95, seed=None):
     """
