+++ permute/tests/test_utils.py
@@ -4,10 +4,25 @@ import numpy as np
 from numpy.random import RandomState
 from numpy.testing import assert_equal
 
-from ..utils import (get_prng,
+from ..utils import (binom_conf_interval,
+                     get_prng,
                      permute_rows,
                      permute_within_groups,
                      permute_incidence_fixed_sums)
+
+
+def test_binom_conf_interval():
+    res = binom_conf_interval(10, 3)
+    expected = (0.05154625578928545, 0.6915018049393984)
+    np.testing.assert_equal(res, expected)
+
+    res2 = binom_conf_interval(10, 5, cl=0.95, alternative="upper")
+    expected2 = (0.0, 0.7775588989918742)
+    np.testing.assert_equal(res2, expected2)
+
+    res3 = binom_conf_interval(10, 5, cl=0.95, alternative="lower")
+    expected3 = (0.22244110100812578, 1.0)
+    np.testing.assert_equal(res3, expected3)
 
 
 def test_get_random_state():
